+++ src/main/java/org/example/Chocolate.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Chocolate extends Product {
-    public Chocolate(String name, int price) {
+    public Chocolate(String name, double price) {
         super(name, "Chocolate", price);
     }
 }
