+++ src/main/java/org/example/Crisps.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Crisps extends Product {
-    public Crisps(String name, int price) {
+    public Crisps(String name, double price) {
         super(name, "crisps", price);
     }
 }
