+++ src/main/java/org/example/Drink.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Drink extends Product{
-    public Drink(String name, int price) {
+    public Drink(String name, double price) {
         super(name,"drink", price);
     }
 }
